+++ src/settings/MappingConfiguration/index.js
@@ -1,0 +1,1 @@
+export { MappingConfiguration } from './MappingConfiguration';