+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## (IN PROGRESS)
 
 * Module setup. Refs UIGS-1.
+* Add GOBI integration area to FOLIO settings. Refs UIGS-2.